Fix dynamic total for previous days
Only add the running timer to the daily total of a row in the task
history UI if the row's date matches the current running task's start
date.

Previously, the running timer would be added based on the index of the
history row. This meant that the first task of the day would incorrectly
increment the previous day's total, as the row for the current day is
only added once the running timer is stopped.